Move log/benchmark to log/internal
